pyproject.toml
[project]
name = "weatherkit-samples"
version = "0.1.0"
description = "Sample code for using Apple WeatherKit REST API with Python"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
